Generate the ModelingCmd enum with a macro
